% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cctd_meso}
\alias{cctd_meso}
\title{California Current Trophic Database}
\format{
\code{cctd_meso} A data frame with 227605 rows and 9 columns:
\describe{
\item{predator_id, predator_common_name, predator_scientific_name}{ID of
individual predator; common name and scientific name of species.}
\item{prey_id, prey_common_name, prey_scientific_name}{ID of individual
prey item; common name and scientific name of species.}
\item{rreas_yoyrockfish, rreas_anchovy}{Log-transformed catch per unit
effort of young-of-the-year rockfish and anchovy, respectively.}
}
}
\source{
\insertRef{Iglesias_2023}{marinecs100b}
}
\usage{
cctd_meso
}
\description{
Predator diet samples from the California Current Trophic Database, with prey
classified as mesopelagic or coastal pelagic.
}
\keyword{datasets}
