% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hothand.R
\name{hothand}
\alias{hothand}
\alias{plot_hothand_simulation}
\alias{plot_hothand_trends}
\title{Plot the results of "hot hand" simulations}
\usage{
plot_hothand_simulation(sim_results)

plot_hothand_trends(trend_results)
}
\arguments{
\item{sim_results}{\verb{[numeric(n)]} A vector with post-streak chomp-rates for
multiple simulations with the same simulation parameters.}

\item{trend_results}{\verb{[data.frame]} A data frame with columns
\code{chomp_attempts}, \code{chomp_rate}, and \code{streak_rate}. Each row represents the
results from a single hot hand simulation. The simulation parameters for
each row are described by columns \code{chomp_attempts} and \code{chomp_rate}.}
}
\value{
\verb{[ggplot]} A ggplot of:
\itemize{
\item \code{plot_hothand_trends()}: the mean "post-streak" chomp rate (solid lines)
over chomp attempts (dashed lines) for different independent chomp rates.
\item \code{plot_hothand_simulation()}: a histogram of "post-streak" chomp rates
with the mean simulated rate represented by a solid red line and the
independent chomp rate indicated by a dashed red line.
}
}
\description{
Plot the results of "hot hand" simulations
}
