% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kefj}
\alias{kefj}
\alias{kefj_temperature}
\alias{kefj_site}
\alias{kefj_datetime}
\alias{kefj_tidelevel}
\alias{kefj_exposure}
\alias{kefj_season}
\title{Kenai Fjords Temperature Readings}
\format{
\verb{kefj_*} are all vectors of length 1295038.

\code{kefj_temperature} A numeric vector with the temperature readings in Celsius.

\code{kefj_site} A character vector with the name of the site (one of Aialik,
Harris, McCarty, Nuka_Bay, or Nuka_Pass) of the temperature reading.

\code{kefj_datetime} A POSIXct vector with the date and time of the temperature
record.

\code{kefj_tidelevel} A numeric vector with the predicted tide level (meters) at
the time of the temperature reading.

\code{kefj_exposure} A character vector indicating the type of temperature
reading:
\describe{
\item{\code{air}}{The predicted tidal elevation was at least 0.5 m lower than the
tide height at which the logger was deployed}
\item{\code{transition}}{The predicted tidal elevation was within 0.5 m of the
tide height at which the logger was deployed}
\item{\code{water}}{The predicted tidal elevation was at least 0.5 m higher than
the tide height at which the logger was deployed}
\item{\code{air/transition}}{The tidal elevation and tide height do not match the
parameters for air or transition temperature and the temperature reading
does not align with water temperature}
}

\code{kefj_season} A five-level factor indicating season, relating to mussel life
history:
\describe{
\item{\verb{Late winter}}{January–March, mussel gonad development}
\item{\code{Spring}}{April–May, spawning and settlement}
\item{\code{Summer}}{June–July, spawning and early growth}
\item{\code{Fall}}{August–October, spawning and late growth}
\item{\verb{Early winter}}{November–December, senescence and early gonad
development}
}
}
\source{
\insertRef{USGS_2024}{marinecs100b}
}
\usage{
kefj_temperature

kefj_site

kefj_datetime

kefj_tidelevel

kefj_exposure

kefj_season
}
\description{
Temperature readings from five sites in Kenai Fjords 2007-2024.
}
\keyword{datasets}
