% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ild.R
\name{plot_ild}
\alias{plot_ild}
\title{Plot the isothermal layer depth}
\usage{
plot_ild(cast, n, h_b, t_b, h_ref, t_ref)
}
\arguments{
\item{cast}{\verb{[data.frame]} A data.frame with the CTD cast data, containing
columns \code{temperature} and \code{depth}.}

\item{n}{\verb{[numeric(1)]} The index of the last CTD point above the base of the
ILD.}

\item{h_b}{\verb{[numeric(1)]} The depth of the ILD base.}

\item{t_b}{\verb{[numeric(1)]} The temperature of the ILD base.}

\item{h_ref}{\verb{[numeric(1)]} The depth of the ILD reference temperature.}

\item{t_ref}{\verb{[numeric(1)]} The ILD reference temperature.}
}
\value{
\verb{[ggplot]} A ggplot of the CTD cast in the isothermal layer, with
landmarks labeled.
}
\description{
Creates a plot of temperature over depth, highlighting the landmarks needed
to calculate isothermal layer depth (ILD) according to the methods of Kara et
al. (2000).
}
\examples{
# You (the student) will calculate these values in one of the modules!
n <- 22
h_b <- 23.83
t_b <- 16.41
h_ref <- 19
t_ref <- 17.21
cast <- calcofi_ctd[calcofi_ctd$station_id == "083.3 070.0", ]
plot_ild(cast, n, h_b, t_b, h_ref, t_ref)
}
\references{
\insertRef{Kara_2000}{marinecs100b}
}
