% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{plot_montyhall}
\alias{plot_montyhall}
\title{Plot the results of two different Monty Hall strategies}
\usage{
plot_montyhall(n_trials, stayput_success, switch_success)
}
\arguments{
\item{n_trials}{\verb{[numeric(1)]} The number of trials simulated.}

\item{stayput_success}{\verb{[numeric(1)]} The number of success from strategy
"stay put".}

\item{switch_success}{\verb{[numeric(1)]} The number of success from strategy
"switch".}
}
\value{
\verb{[ggplot]} A ggplot of a bar plot showing the differences in success
between the two strategies.
}
\description{
Plot the results of two different Monty Hall strategies
}
\examples{
# Note: your results will look very different if you simulated the two
# strategies correctly!
plot_montyhall(1000, 500, 500)
}
