% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rreas_comu}
\alias{rreas_comu}
\title{RREAS survey and Commun Murre diet data}
\format{
\code{rreas_comu} A data frame with 37 rows and 5 columns:
\describe{
\item{year}{Year of data}
\item{comu_yoyrockfish, comu_anchovy}{Fraction of Common Murre diet
composed of young-of-the-year rockfish and anchovy, respectively.}
\item{rreas_yoyrockfish, rreas_anchovy}{Log-transformed catch per unit
effort of young-of-the-year rockfish and anchovy, respectively.}
}
}
\source{
\insertRef{Santora_2021}{marinecs100b}
}
\usage{
rreas_comu
}
\description{
Two independent proxies for forage fish abundance in the California Current.
(1) Log-transformed estimates of catch per unit effort of young-of-the-year
rockfish and anchovy from the Rockfish Recruitment Ecosystem Assessment
Survey. (2) Percentage of Common Murre diet composed of the same two species
observed on Southeast Farallon Island. Rockfish data spans 1983-2019, anchovy
1990-2019.
}
\keyword{datasets}
