% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woa_sal}
\alias{woa_sal}
\title{World Ocean Atlas data (salinity)}
\format{
\code{woa_long} A data frame with 2810255 rows and 4 columns:
\describe{
\item{latitude, longitude}{Latitude and longitude in decimal degrees. Longitude in the range -180, 180.}
\item{salinity}{Salinity (unitless).}
}
}
\source{
\url{https://www.ncei.noaa.gov/access/world-ocean-atlas-2023/bin/woa23.pl}
}
\usage{
woa_sal
}
\description{
The latest World Ocean Atlas data (as of 2025). Specifically, the statistical
mean of salinity for the "averaged decades" climatology at the annual time
scale. Only surface salinity is included
}
\keyword{datasets}
